Rename tests cases
Renamed test cases named after HTTP verbs.

- The PUT tests was testing more various facts functionalities than
  just the PUT method.
- Other test cases in fact tests various host related functionality.
  Renamed to reflect that.
