+++ tests/test_api_hosts_delete.py
