+++ tests/test_api_hosts_update.py
